refactor : SocialInfo 패키지 이동
